docs/requirements.txt
numpy
sphinx-rtd-theme
nbsphinx
sphinxcontrib-bibtex

